requirements.txt
keras==3.5.0
numpy==1.26.4
pillow==10.4.0
tensorflow==2.17.0
tensorflow-datasets==4.9.6
streamlit_drawable_canvas==0.9.3
streamlit==1.38.0
